Refactor upload modal, set up correct keys for FormData
