+++ app/javascript/components/modals/upload-video/elements/index.ts
@@ -1,3 +1,9 @@
 export { ExerciseTrackIndicator } from './ExerciseTrackIndicator'
 export { UploadVideoTextInput } from './UploadVideoTextInput'
 export { UploadVideoForm } from './UploadVideoForm'
+export {
+  RetrieveVideoForm,
+  VideoDataResponse,
+  CommunityVideo,
+} from './RetrieveVideoForm'
+export { UploadVideoModalHeader } from './UploadVideoModalHeader'
